# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elizabeth Okwori's Confectionery NGO Enrollment Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  sex?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  createdAt: string;
}

export interface UserProfileInput {
  fullName?: string;
  phone?: string;
  sex?: string;
  profilePhotoUrl?: string;
}

export interface RegistrationInput {
  role: string;
  phone: string;
  sex: string;
  /** @nullable */
  officerCode?: string | null;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface SkillInput {
  name: string;
  category: string;
  description?: string;
}

export interface Package {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  duration: string;
  createdAt: string;
}

export interface PackageInput {
  name: string;
  description?: string;
  price: number;
  duration: string;
}

export interface Enrollment {
  id: number;
  userId: number;
  packageId: number;
  status: string;
  skillIds: number[];
  /** @nullable */
  whatsappNumber?: string | null;
  createdAt: string;
  user?: UserProfile;
  package?: Package;
  skills?: Skill[];
}

export interface EnrollmentInput {
  packageId: number;
  skillIds: number[];
}

export interface Payment {
  id: number;
  enrollmentId: number;
  userId: number;
  proofUrl: string;
  fileName: string;
  status: string;
  uploadedAt: string;
  /** @nullable */
  reviewedAt?: string | null;
  user?: UserProfile;
  enrollment?: Enrollment;
}

export interface PaymentInput {
  enrollmentId: number;
  proofUrl: string;
  fileName: string;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationInput {
  /** @nullable */
  targetUserId?: number | null;
  title: string;
  message: string;
  type: string;
}

export type DashboardStatsEnrollmentsByStatusItem = {
  status: string;
  count: number;
};

export interface DashboardStats {
  totalParticipants: number;
  totalOfficers: number;
  pendingPayments: number;
  approvedEnrollments: number;
  activeSkills: number;
  recentNotifications: number;
  enrollmentsByStatus: DashboardStatsEnrollmentsByStatusItem[];
  recentActivity: Notification[];
}

export interface PortalSettings {
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  officerCode?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  address?: string | null;
}

export interface PortalSettingsInput {
  whatsappNumber?: string;
  contactEmail?: string;
  address?: string;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  createdAt: string;
  createdBy: number;
}

export interface AnnouncementInput {
  title: string;
  content: string;
}

export type ListParticipantsParams = {
search?: string;
role?: string;
};

export type ListEnrollmentsParams = {
status?: string;
};