# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elizabeth Okwori's Confectionery NGO Enrollment Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List enrollment packages
      responses:
        "200":
          description: Packages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package (officer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    put:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package (officer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package (officer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments (officer gets all, participant gets own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Enrollments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Create enrollment (participant only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Created enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    get:
      operationId: getEnrollment
      tags: [enrollments]
      summary: Get enrollment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrollment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}/abort:
    patch:
      operationId: abortEnrollment
      tags: [enrollments]
      summary: Abort an enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Aborted enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send notification (officer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - fullName
        - email
        - createdAt
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - createdAt
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: integer
        duration:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - duration
        - createdAt
    PackageInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        duration:
          type: string
      required:
        - name
        - price
        - duration
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        packageId:
          type: integer
        status:
          type: string
        skillIds:
          type: array
          items:
            type: integer
        whatsappNumber:
          type: ["string", "null"]
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        package:
          $ref: "#/components/schemas/Package"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
      required:
        - id
        - userId
        - packageId
        - status
        - skillIds
        - createdAt
    EnrollmentInput:
      type: object
      properties:
        packageId:
          type: integer
        skillIds:
          type: array
          items:
            type: integer
      required:
        - packageId
        - skillIds
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
    NotificationInput:
      type: object
      properties:
        targetUserId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
        type:
          type: string
      required:
        - title
        - message
        - type
...