# lib/api-zod/src/generated/types/packageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elizabeth Okwori's Confectionery NGO Enrollment Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageInput {
  name: string;
  description?: string;
  price: number;
  duration: string;
}